+++ envdata/admin.py
@@ -1,7 +1,7 @@
 
 
 from django.contrib import admin
-from envdata.models import (Company, Fuel, Sf6, Refrigerant, Energy, Travel, Waste)
+from envdata.models import (Company, Fuel, Sf6, Refrigerant, Energy, Travel, Waste,Target)
 
 # Register your models here.
 admin.site.register(Company)
@@ -11,3 +11,4 @@ admin.site.register(Refrigerant)
 admin.site.register(Energy)
 admin.site.register(Travel)
 admin.site.register(Waste)
+admin.site.register(Target)