+++ envdata/views/__init__.py
@@ -6,3 +6,4 @@ from .sf6_views import *
 from .travel_views import *
 from .waste_views import *
 from .navbar_views import *
+from .target_views import *